Add login/logout button
